add embedded search to the correct place
