+++ frontend/src/styles/main.css
@@ -86,6 +86,12 @@ body {
 	border-radius: 4px;
 }
 
+.embedded-search-input{
+	border: none!important;
+    background: transparent!important;
+    border-bottom: 0.5px solid #868E96!important;
+}
+
 /*turn off the shadow box on search*/
 /*TODO pass in outer div id*/
 .input-group{
@@ -95,6 +101,6 @@ body {
 .filter-select{
 	border: none!important;
     background: transparent!important;
-    border-bottom: 1px solid grey!important;
+    border-bottom: 1px solid #868E96!important;
     font-weight: bold;
 }
